Added support for multi threading. Fixed issues with static JsonParsers. Made HandshakeUtil methods static. Renamed HandshakeServerTest to HandshakeUtilTest to better reflect it's purpose.
