+++ hnclient-v2/src/main/resources/application.properties
@@ -9,6 +9,7 @@ port = 8080
 server-socket=5555
 socket-read-sleep-time=100
 max-socket-read-tries=100
+thread-pool-size=5
 
 ## to hnsecure
 # For HNSecure:
